Put the SearchBar component and its subcomponents into a separate subfolder
