+++ src/components/searchBar/AdvancedSearchOption.js
@@ -8,7 +8,7 @@ const AdvancedSearchOption = (props) => {
             <input type = {props.type} />
             <div className = {Styles.advancedSearchSelectables}>
                 <AdvancedSearchSelectable type = {props.type} name = {props.name} value = "Exact" />
-                {props.type === "date" &&
+                {props.type === "date" &&   // Date options also get Before... and After... switches, everything else is just toggled between Exact and not
                     <>
                         <AdvancedSearchSelectable type = "date" name = {props.name} value = "Before" />
                         <AdvancedSearchSelectable type = "date" name = {props.name} value = "After" />
