+++ src/components/searchBar/AdvancedSearchOption.module.scss
