+++ src/components/searchBar/SearchBar.module.scss
