+++ src/components/searchBar/SearchBarAdvanced.module.scss
